Fix datatype in listener_test.
Summary:
.

Test Plan:
A new test EventListenerTest::OnWALOperationTest has been added. The
test verifies that during restore the sequential file reader is called
and the notifications are fired.

Reviewers:
yanqin, rmanji

Subscribers:

Tasks:
T75121251

Tags:

+++ db/listener_test.cc
@@ -1209,7 +1209,7 @@ TEST_F(EventListenerTest, OnWALOperationTest) {
   ASSERT_OK(dbfull()->TEST_WaitForFlushMemTable());
   Close();
 
-  unsigned long seq_reads = listener->file_seq_reads_.load();
+  size_t seq_reads = listener->file_seq_reads_.load();
   Reopen(options);
   ASSERT_GT(listener->file_seq_reads_.load(), seq_reads);
 }
